+++ bank-feeds/src/sdk/models/shared/banktransactions.ts
@@ -5,7 +5,31 @@
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { Expose } from "class-transformer";
 
-export class CreateBankTransaction extends SpeakeasyBase {
+/**
+ * Type of transaction for the bank statement line.
+ */
+export enum BankTransactionsBankTransactionType {
+    Unknown = "Unknown",
+    Credit = "Credit",
+    Debit = "Debit",
+    Int = "Int",
+    Div = "Div",
+    Fee = "Fee",
+    SerChg = "SerChg",
+    Dep = "Dep",
+    Atm = "Atm",
+    Pos = "Pos",
+    Xfer = "Xfer",
+    Check = "Check",
+    Payment = "Payment",
+    Cash = "Cash",
+    DirectDep = "DirectDep",
+    DirectDebit = "DirectDebit",
+    RepeatPmt = "RepeatPmt",
+    Other = "Other",
+}
+
+export class BankTransactions extends SpeakeasyBase {
     /**
      * The amount transacted in the bank transaction.
      */
@@ -44,8 +68,15 @@ export class CreateBankTransaction extends SpeakeasyBase {
      * > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
      */
     @SpeakeasyMetadata()
-    @Expose({ name: "date" })
-    date?: string;
+    @Expose({ name: "clearedOnDate" })
+    clearedOnDate?: string;
+
+    /**
+     * The giving or receiving party such as a person or organization.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "counterparty" })
+    counterparty?: string;
 
     /**
      * Description of the bank transaction.
@@ -60,4 +91,25 @@ export class CreateBankTransaction extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "id" })
     id?: string;
+
+    /**
+     * `True` if the bank transaction has been [reconciled](https://www.xero.com/uk/guides/what-is-bank-reconciliation/) in the accounting platform.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "reconciled" })
+    reconciled?: boolean;
+
+    /**
+     * An optional reference to the bank transaction.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "reference" })
+    reference?: string;
+
+    /**
+     * Type of transaction for the bank statement line.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "transactionType" })
+    transactionType?: BankTransactionsBankTransactionType;
 }