+++ bank-feeds/RELEASES.md
@@ -677,3 +677,13 @@ Based on:
 - [typescript v2.2.1] bank-feeds
 ### Releases
 - [NPM v2.2.1] https://www.npmjs.com/package/@codat/bank-feeds/v/2.2.1 - bank-feeds
+
+## 2023-09-28 00:05:25
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.91.2 (2.131.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.2.2] bank-feeds
+### Releases
+- [NPM v2.2.2] https://www.npmjs.com/package/@codat/bank-feeds/v/2.2.2 - bank-feeds